requirements.txt
torch
torchvision
torchaudio
scikit-learn
numpy
pandas
matplotlib
tqdm
jupyter
uvicorn
timm
pillow
fastapi[standard]